pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mirai"
version = "0.1.0"
description = "AI-powered personal assistant"
authors = [ "abdelmalek0 <djamaa.abdelmalek98@gmail.com>",]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
langchain-groq = "^0.2.1"
langchain = "^0.3.7"
python-dotenv = "^1.0.1"
reorder-python-imports = "^3.14.0"
langchain-community = "^0.3.7"
langgraph = "^0.2.50"
elevenlabs = "^1.12.1"
pydub = "^0.25.1"
